Reorganize v2 insights
