+++ src/backend/common/helpers/insights_v2/leaderboards/calculator.py
@@ -1,24 +1,17 @@
-from abc import ABC, abstractmethod
+from abc import abstractmethod
 from collections import defaultdict
-from typing import Any, DefaultDict, Dict, List, Optional, Tuple
+from typing import DefaultDict, Dict, List, Optional, Tuple
 
-from google.appengine.ext import ndb
-
-from backend.common.consts.event_type import SEASON_EVENT_TYPES
-from backend.common.consts.renamed_districts import RenamedDistricts
-from backend.common.helpers.season_helper import SeasonHelper
-from backend.common.models.event import Event
+from backend.common.helpers.insights_v2.base import InsightV2Calculator
+from backend.common.helpers.insights_v2.names import InsightV2NameEntry
 from backend.common.models.insight_v2 import (
     InsightCategory,
     InsightV2,
-    InsightV2NameEntry,
     LeaderboardDataV2,
     LeaderboardKeyType,
     LeaderboardRankingV2,
 )
 from backend.common.models.keys import Year
-from backend.common.queries.district_query import AllDistrictTeamsQuery
-from backend.common.queries.event_query import EventListQuery
 
 LEADERBOARD_TOP_N = 25
 LEADERBOARD_MAX_KEYS_PER_RANKING = 10
@@ -71,37 +64,7 @@ def build_leaderboard_pair_rankings(
     ][:top_n]
 
 
-def _build_team_district_map() -> Dict[str, str]:
-    """
-    Returns {team_key: district_abbrev} for all teams, using each team's most
-    recent district membership.
-    """
-    all_district_teams = AllDistrictTeamsQuery().fetch()
-
-    team_district_teams: Dict[str, List[Any]] = defaultdict(list)
-    for dt in all_district_teams:
-        if dt.team:
-            team_district_teams[str(dt.team.id())].append(dt)
-
-    return {
-        team_key: RenamedDistricts.get_latest_code(
-            str(max(dts, key=lambda dt: dt.year).district_key.id())[4:]
-        )
-        for team_key, dts in team_district_teams.items()
-    }
-
-
-class InsightV2Calculator(ABC):
-    @abstractmethod
-    def on_event(self, event: Event) -> None: ...
-
-    @abstractmethod
-    def make_insights(
-        self, year: Year, team_to_district: Dict[str, str]
-    ) -> List[InsightV2]: ...
-
-
-class LeaderboardV2Calculator(InsightV2Calculator, ABC):
+class LeaderboardV2Calculator(InsightV2Calculator):
     def __init__(self) -> None:
         self.counts: Dict[str, int] = defaultdict(int)
 
@@ -188,33 +151,3 @@ class LeaderboardV2Calculator(InsightV2Calculator, ABC):
             )
 
         return insights
-
-
-def compute_insights_for_year(
-    year: Year, calculators: List[InsightV2Calculator]
-) -> List[InsightV2]:
-    """
-    Iterates over all season events for a year (or all years if year=0),
-    calling on_event() on every calculator once per event. Preps and clears
-    event relations per-event so memory stays bounded.
-    """
-    event_years = [year] if year != 0 else SeasonHelper.get_valid_years()
-    for event_year in event_years:
-        for event in EventListQuery(year=event_year).fetch():
-            if event.event_type_enum not in SEASON_EVENT_TYPES:
-                continue
-
-            event.prep_awards()
-            event.prep_matches()
-            for calc in calculators:
-                calc.on_event(event)
-            event.clear_awards()
-            event.clear_matches()
-            ndb.get_context().clear_cache()
-
-    team_to_district = _build_team_district_map()
-
-    insights = []
-    for calc in calculators:
-        insights.extend(calc.make_insights(year, team_to_district))
-    return insights
