+++ src/backend/common/helpers/tests/insights_v2/leaderboards/blue_banners_test.py
@@ -1,10 +1,12 @@
-from backend.common.helpers.insights_v2.blue_banners import BlueBannersV2Calculator
-from backend.common.helpers.insights_v2.compute import compute_insights_for_year
+from backend.common.helpers.insights_v2.leaderboards.blue_banners import (
+    BlueBannersV2Calculator,
+)
+from backend.common.helpers.insights_v2.registry import compute_insights_for_year
 
 
 def test_blue_banners_year(ndb_stub, test_data_importer) -> None:
-    test_data_importer.import_event(__file__, "../data/2024nytr.json")
-    test_data_importer.import_award_list(__file__, "../data/2024nytr_awards.json")
+    test_data_importer.import_event(__file__, "../../data/2024nytr.json")
+    test_data_importer.import_award_list(__file__, "../../data/2024nytr_awards.json")
 
     insights = compute_insights_for_year(2024, [BlueBannersV2Calculator()])
 
@@ -17,10 +19,10 @@ def test_blue_banners_year(ndb_stub, test_data_importer) -> None:
 
 
 def test_blue_banners_overall(ndb_stub, test_data_importer) -> None:
-    test_data_importer.import_event(__file__, "../data/2019nyny.json")
-    test_data_importer.import_award_list(__file__, "../data/2019nyny_awards.json")
-    test_data_importer.import_event(__file__, "../data/2024nytr.json")
-    test_data_importer.import_award_list(__file__, "../data/2024nytr_awards.json")
+    test_data_importer.import_event(__file__, "../../data/2019nyny.json")
+    test_data_importer.import_award_list(__file__, "../../data/2019nyny_awards.json")
+    test_data_importer.import_event(__file__, "../../data/2024nytr.json")
+    test_data_importer.import_award_list(__file__, "../../data/2024nytr_awards.json")
 
     insights = compute_insights_for_year(0, [BlueBannersV2Calculator()])
 
@@ -29,8 +31,8 @@ def test_blue_banners_overall(ndb_stub, test_data_importer) -> None:
 
 
 def test_key_name(ndb_stub, test_data_importer) -> None:
-    test_data_importer.import_event(__file__, "../data/2024nytr.json")
-    test_data_importer.import_award_list(__file__, "../data/2024nytr_awards.json")
+    test_data_importer.import_event(__file__, "../../data/2024nytr.json")
+    test_data_importer.import_award_list(__file__, "../../data/2024nytr_awards.json")
 
     insights = compute_insights_for_year(2024, [BlueBannersV2Calculator()])
     assert insights[0].key_name == "2024_v2_leaderboard_blue_banners"
